+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `DisplayOnlyUrlType` is ready to use
     - `LocalizedFullWidthType` is ready to use
     - `ProductParameterValueType` is ready to use
+    - `ProductCalculatedPricesType` is ready to use
 
 #### Changed
 - [#385 - AccessDeniedHttpException replaced by AccessDeniedException](https://github.com/shopsys/shopsys/pull/385)